chore : refactoring, move package, increase readability
