+++ src/main/kotlin/json/delivery/socialnetworkservice/app/api/v1/FavoriteController.kt
@@ -24,6 +24,6 @@ class FavoriteController(
         @PathVariable articleId: String,
         @Validated @RequestBody request: FavoriteRequest,
     ): FavoriteResponse {
-        return favoriteUsecase.execute(UserId(request.userId), articleId).let { FavoriteResponse(it) };
+        return FavoriteResponse(favoriteUsecase.execute(UserId(request.userId), articleId));
     }
 }
