+++ src/main/kotlin/json/delivery/socialnetworkservice/app/application/FavoriteService.kt
@@ -1,5 +1,6 @@
 package json.delivery.socialnetworkservice.app.application
 
+import json.delivery.socialnetworkservice.app.application.dto.FavoriteUsecaseOutputDto
 import json.delivery.socialnetworkservice.app.domain.UserId
 import json.delivery.socialnetworkservice.app.infrastructure.ArticleRepository
 import org.springframework.stereotype.Service
