+++ src/main/kotlin/json/delivery/socialnetworkservice/app/application/FavoriteUsecase.kt
@@ -1,5 +1,6 @@
 package json.delivery.socialnetworkservice.app.application
 
+import json.delivery.socialnetworkservice.app.application.dto.FavoriteUsecaseOutputDto
 import json.delivery.socialnetworkservice.app.domain.UserId
 
 interface FavoriteUsecase {
