+++ src/main/kotlin/json/delivery/socialnetworkservice/app/application/dto/FavoriteUsecaseOutputDto.kt
@@ -1,0 +1,8 @@
+package json.delivery.socialnetworkservice.app.application.dto
+
+import json.delivery.socialnetworkservice.app.domain.UserId
+
+class FavoriteUsecaseOutputDto(
+    val userId: UserId,
+    val articleId: String,
+)
